etl/requirements_simple.txt
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
